Rename core/data_schemas.py to core/schemas.py
