+++ rag/core/interfaces.py
@@ -1,7 +1,7 @@
 from typing import Protocol, runtime_checkable
 from pathlib import Path
 
-from rag.core.data_schemas import ParseResult, BookEntry
+from rag.core.schemas import ParseResult, BookEntry
 
 
 @runtime_checkable
